refactor(memtable): dedup table proxy code
