+++ ibis/backends/base/sql/compiler/select_builder.py
@@ -301,12 +301,7 @@ class SelectBuilder:
             self.table_set = table
             self.filters = filters
 
-    def _collect_PandasInMemoryTable(self, node, toplevel=False):
-        if toplevel:
-            self.select_set = [node]
-            self.table_set = node
-
-    def _collect_PyArrowInMemoryTable(self, node, toplevel=False):
+    def _collect_InMemoryTable(self, node, toplevel=False):
         if toplevel:
             self.select_set = [node]
             self.table_set = node
